requirements.txt
python-telegram-bot[job-queue]
discord.py
asyncpg
httpx
gunicorn
aiohttp
